requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
python-telegram-bot==21.9
selenium==4.27.1
